No item UI Implementation done. Some Animation also implemented.

+++ Views/ListView.swift
@@ -13,18 +13,26 @@ struct ListView: View {
     @State var reloadFlag : Bool = false
     
     var body: some View {
-        List{
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                            reloadFlag.toggle()
-                        }
+        ZStack{
+            if self.listViewModel.items.isEmpty {
+                NoItemView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            }else{
+                List{
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                    reloadFlag.toggle()
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(PlainListStyle())
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
         .navigationTitle("ToDo List 📝")
         .navigationBarItems(leading: EditButton(), trailing: NavigationLink("Add", destination: AddView()))
